#29 231023-2

+++ Jin/Jin/BtnUI.cpp
@@ -36,7 +36,11 @@ void BtnUI::render(HDC _dc)
 	}
 	else
 	{
-		Rectangle(_dc, vPos.x, vPos.y, vPos.x + vScale.x, vPos.y + vScale.y);
+		Rectangle(_dc
+			, (int)(vPos.x)
+			, (int)(vPos.y)
+			, (int)(vPos.x + vScale.x)
+			, (int)(vPos.y + vScale.y));
 	}
 	Super::render(_dc);
 }
