+++ Jin/Jin/EditorLevel.cpp
@@ -75,11 +75,11 @@ void EditorLevel::tick()
 		Vec2 vMousePos = KeyMgr::GetInst()->GetMousePos();
 		vMousePos = Camera::GetInst()->GerRealPos(vMousePos);
 
-		int col = vMousePos.x / TILE_SIZE;
-		int row = vMousePos.y / TILE_SIZE;
+		int col = (int)vMousePos.x / TILE_SIZE;
+		int row = (int)vMousePos.y / TILE_SIZE;
 		int idx = GetTileCol() * row + col;
 
-		if (!(GetTileCol() <= col) && !(GetTileCol() <= row)
+		if (!((int)GetTileCol() <= col) && !((int)GetTileCol() <= row)
 			 && !(vMousePos.x < 0.f) && !(vMousePos.y < 0.f) )
 		{
 			const vector<Obj*>& vecTiles = GetLayer(LAYER::TILE)->GetObjects();
